change int() to float()

+++ features/extractor.py
@@ -66,7 +66,10 @@ class Outcome(FeatureExtractor):
         if score in Outcome.LABEL2INT:
             return ("Outcome", Outcome.LABEL2INT[score])
         else:
-            return ("Outcome", int(answer.contentScore))
+            return ("Outcome", float(answer.contentScore))
+            #todo:changed this to float()   File "/home/akrnshva/git/isaac-ml-service/features/extractor.py", line 69, in extract
+            # return ("Outcome", int(answer.contentScore))
+            # ValueError: invalid literal for int() with base 10: '-98.0'
 
 class Diagnosis(FeatureExtractor):
